+++ Source/Applications/Desktop/Vodovoz.Presentation.ViewModels/Controls/EntitySelection/SelectionDialogSettings.cs
@@ -11,6 +11,5 @@
 		public int WindowWidth { get; set; } = 220;
 		public int ButtonHeight { get; set; } = 30;
 		public int ButtonWidth { get; set; } = 50;
-
 	}
 }
